Network.CGI.Compat

Description

Compatibility functions for the old Network.CGI API.

Documentation

data Html

wrapper :: ([(String, String)] -> IO Html) -> IO ()

Deprecated: Use the new interface.

Compatibility wrapper for the old CGI interface. Output the output from a function from CGI environment and input variables to an HTML document.

pwrapper

Arguments

:: PortID
The port to run the server on.

-> ([(String, String)] -> IO Html)

-> IO ()

Deprecated: Use the new interface.

Compatibility wrapper for the old CGI interface. Runs a simple CGI server. Note: if using Windows, you might need to wrap withSocketsDo around main.

connectToCGIScript :: String -> PortID -> IO ()

Deprecated: Use the new interface.

Note: if using Windows, you might need to wrap withSocketsDo around main.